lab/app/requirements.txt
Flask==3.1.0
requests==2.32.3
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-flask==0.52b1
opentelemetry-instrumentation-requests==0.52b1
